Added mxdriverinfo options for automatically generating tables for the LaTeX-based MX Driver Reference Manual.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@714 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/d_epix_xclib.h
@@ -34,7 +34,7 @@ typedef struct {
 	double default_trigger_time;
 	long pixel_clock_divisor;
 
-	size_t num_write_test_array_bytes;
+	unsigned long num_write_test_array_bytes;
 	uint16_t *write_test_array;
 
 } MX_EPIX_XCLIB_VIDEO_INPUT;
